pyproject.toml
[project]
name = "scrape"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "playwright>=1.58.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "transformers>=5.1.0",
    "opencv-python>=4.13.0.92",
    "numpy>=2.4.2",
    "pillow>=12.1.1",
]
